requirements/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
lxml==5.3.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pip-review==1.3.0
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
review==1510
six==1.16.0
soupsieve==2.6
style==1.1.6
tomli==2.1.0
tzdata==2024.2
update==0.0.1
urllib3==2.2.3
webdriver-manager==4.0.2
